% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_elections}
\alias{add_elections}
\title{Add Election Dates to Dated Datasets}
\usage{
add_elections(
  data = NULL,
  date = "date",
  last_name = "last_elec",
  next_name = "next_elec",
  which = "both"
)
}
\arguments{
\item{data}{A dataset that includes a date variable.}

\item{date}{The name of your date variable in your data.}

\item{last_name}{What to call the new column of last election dates. Defaults to "last_elec".}

\item{next_name}{What to call the new column of next election dates. Defaults to "next_elec".}

\item{which}{Which elections to return: "last", "next", or "both". Defaults to "both".}
}
\value{
A tibble of data.
}
\description{
It's often useful to know when elections occurred relative to the cases in your data. This function takes a dataset that includes a dated variable and adds to it the date of the last election, next election, or both.
}
\examples{
add_elections(data = pollbasepro, date = "date", which = "both")
}
