% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_lead}
\alias{add_lead}
\title{Add Lead of One Party Over Another}
\usage{
add_lead(data = NULL, party1 = "con", party2 = "lab", name = "lead")
}
\arguments{
\item{data}{A dataset containing variables called {{party}}_est and {{party}}_err}

\item{party1}{The name of party}

\item{party2}{The name of your date variable in your data.}

\item{name}{The name . Defaults to "lead".}
}
\value{
A tibble of data.
}
\description{
We often want to know how much further one party is ahead of another in the polls. But this can be tricky when the polls contain error. This function takes a dataset with columns named "{{party}}_est" and "{{party}}_err", computes the lead of one party over another, and appends it plus the error in the estimate to the data.
}
\examples{
add_lead(data = pollbasepro, party1 = "con", party2 = "lab")
}
