% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_pm}
\alias{add_pm}
\title{Add Prime Minister to Dated Datasets}
\usage{
add_pm(data = NULL, date = "date", name = "prime_minister")
}
\arguments{
\item{data}{A dataset that includes a date variable.}

\item{date}{The name of your date variable in your data.}

\item{name}{What to call the new column of Prime Ministers. Defaults to "prime_minister".}
}
\value{
A tibble of data.
}
\description{
It's often useful to know who the Prime Minister was at any point in time when analysing political data. This function takes a dataset that includes a dated variable and adds to it a new column that shows who the Prime Minister was at each moment in time.
}
\examples{
dta <- data.frame(date = seq.Date(as.Date("2000-01-01"), as.Date("2020-01-01"), by = "year"))
add_pm(data = dta, date = "date", name = "prime_minister")
}
