% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_pm_party}
\alias{add_pm_party}
\title{Add Prime Minister's Party to Dated Datasets}
\usage{
add_pm_party(data = NULL, date = "date", name = "pm_party")
}
\arguments{
\item{data}{A dataset that includes a date variable.}

\item{date}{The name of your date variable in your data.}

\item{name}{What to call the new column of Prime Ministers. Defaults to "pm_party".}
}
\value{
A tibble of data.
}
\description{
It's often useful to know which party the Prime Minister belonged to at any point in time when analysing political data. This function takes a dataset that includes a dated variable and adds to it a new column that shows the party of the Prime Minister at each moment in time.
}
\examples{
dta <- data.frame(date = seq.Date(as.Date("2000-01-01"), as.Date("2020-01-01"), by = "year"))
add_pm_party(data = dta, date = "date", name = "pm_party")
}
