% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barnett.R
\name{barnett}
\alias{barnett}
\title{Barnett Formula}
\usage{
barnett(eng_fund, pop_prop, comp_fct)
}
\arguments{
\item{eng_fund}{Amount of extra funding allocated to England.}

\item{pop_prop}{Population proportion compared to England.}

\item{comp_fct}{Comparability factor as a proportion. Value reflects how comparable the English departmental programme is to the devolved administration's programme.}
}
\value{
A vector of numeric data
}
\description{
The British government uses the Barnett formula to allocate money for public expenditure in Scotland, Wales, and Northern Ireland based on changes to funding in England (and something other nations too). This function takes the necessary inputs and returns the expected change in funding expected under the Barnett formula.
}
\examples{
barnett(eng_fund = 1e+09, pop_prop = .1034, comp_fct = .997)
}
