% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_pcon_names.R
\name{clean_pcon_names}
\alias{clean_pcon_names}
\title{Clean British Parliamentary Constituency Names}
\usage{
clean_pcon_names(x, year = NULL)
}
\arguments{
\item{x}{A vector of constituency names.}

\item{year}{A specific year to consider names for. Defaults to NULL.}
}
\value{
A vector of cleaned constituency names.
}
\description{
British parliamentary constituencies are not often coded in a consistent manner. Fixing this can also be very time-consuming. This function uses the constituencies dataset in the package to standardise a vector of constituencies names. Note that where strings are uncertain this function can take a while to run.
}
\examples{
clean_pcon_names(c("Ynys Môn", "Derby North", "North, Derby"), year = 2019)
}
