% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{constituency_results}
\alias{constituency_results}
\title{Historic Election Results by Constituency, 1929 onwards}
\format{
A data frame:
\describe{
\item{date}{Election date}
\item{seats}{Number of seats in constituency}
\item{constituency}{Name of constituency}
\item{pano}{Press Association code}
\item{nomis}{NOMIS code}
\item{pca}{PCA code}
\item{ons}{ONS GSS code}
\item{county}{County}
\item{region}{Region or country}
\item{electorate}{Size of constituency electorate}
\item{con}{Number of Conservative votes}
\item{lab}{Number of Labour votes}
\item{lib}{Number of Liberal votes}
\item{nat}{Number of Nationalist votes in Scotland and Wales}
\item{oth}{Number of votes for any other party}
}
}
\usage{
constituency_results
}
\description{
A dataset containing historic constituency-level election results in Britain. Taken from the House of Commons library https://commonslibrary.parliament.uk/research-briefings/cbp-8647/.
}
\keyword{datasets}
