% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_twitter_content}
\alias{create_twitter_content}
\title{Create Twitter images and text}
\usage{
create_twitter_content(data = britpol::pollbasepro, path = NULL)
}
\arguments{
\item{data}{Instance of pollbasepro to use to create your output.}

\item{path}{The file path to where you would like to save your content.}
}
\description{
This function calls the latest pollbasepro data and uses it to create and save social media content.
}
