% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{election_dates}
\alias{election_dates}
\title{List of British Election Dates Since Universal Suffrage}
\format{
A data frame:
\describe{
\item{date}{Date of election}
}
}
\usage{
election_dates
}
\description{
A dataset containing the dates of all British General Elections since the advent of universal suffrage in 1928.
}
\keyword{datasets}
