% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_2015_polls}
\alias{get_2015_polls}
\title{Get Voting Intention Polls, 2010-2015}
\usage{
get_2015_polls()
}
\description{
This function downloads and tidies polling data for the 2010 to 2015 period from Wikipedia.
}
