% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_2017_polls}
\alias{get_2017_polls}
\title{Get Voting Intention Polls, 2015-2017}
\usage{
get_2017_polls()
}
\description{
This function downloads and tidies polling data for the 2015 to 2017 period from Wikipedia.
}
