% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_2019_polls}
\alias{get_2019_polls}
\title{Get Voting Intention Polls, 2017-2019}
\usage{
get_2019_polls()
}
\description{
This function downloads and tidies polling data for the 2017 to 2019 period from Wikipedia.
}
