% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_new_polls}
\alias{get_new_polls}
\title{Get Voting Intention Polls, 2019-Present}
\usage{
get_new_polls()
}
\description{
This function downloads and tidies polling data for the 2019-Present period from Wikipedia.
}
