% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{house_effects}
\alias{house_effects}
\title{Estimated House Effects by Pollster Across All Post-1955 Elections}
\format{
A data frame:
\describe{
\item{date}{Election date}
\item{pollster}{Polling company}
\item{con_est}{Posterior mean: Conservative house effect}
\item{con_err}{Posterior error: Conservative house effect}
\item{lab_est}{Posterior mean: Labour house effect}
\item{lab_err}{Posterior error: Labour house effect}
\item{lib_est}{Posterior mean: Liberal house effect}
\item{lib_err}{Posterior error: Liberal house effect}
}
}
\usage{
house_effects
}
\description{
A dataset containing estimates of historic "house effects" for each polling firm for each party and each election.
}
\keyword{datasets}
