% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{in_text}
\alias{in_text}
\title{In-Text Summary Statistics}
\usage{
in_text(x, text = NULL, suffix = NULL, inside = T, probs = 0.95, digits = 1)
}
\arguments{
\item{x}{The distribution to summarise}

\item{text}{A character string that follows and describes the point estimate.}

\item{suffix}{A character string that follows each estimate.}

\item{inside}{Should the median be inside or outside the brackets? Defaults to T.}

\item{probs}{The level at which to compute the lower and upper bounds. Defaults to 0.95.}

\item{digits}{How many digits to round the summary statistics to. Defaults to 1.}
}
\value{
A text string.
}
\description{
Generate in-text summary statistics for posterior distributions from Bayesian models. Most useful when used in Rmarkdown documents.
}
