% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_redwall.R
\name{is_red_wall}
\alias{is_red_wall}
\title{Is a Constituency in the "Red Wall"?}
\usage{
is_red_wall(x = NULL)
}
\arguments{
\item{x}{A vector of constituency data.}
}
\value{
A boolean vector.
}
\description{
This function takes a vector of constituency names and returns a variable that indicates if each element is in the "Red Wall" or not. These constituencies are coded according to the classification created by James Kanagasooriam.
}
\examples{
\dontrun{
is_redwall(x = "Bury South")
}
}
