% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{lookup_interval}
\alias{lookup_interval}
\title{Lookup Column Values Based on Date Intervals}
\usage{
lookup_interval(lookup_date, interval_data, variable)
}
\arguments{
\item{lookup_date}{A column of dated data.}

\item{interval_data}{A dataset including columns called "start" and "end".}

\item{variable}{The variable to select values from in the interval_data.}
}
\value{
A tibble of data.
}
\description{
Sometimes we need to lookup a value at a particular time based on a date range in another dataset. This function provides a wrapper to data.table to make this quick and easy.
}
