% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mae}
\alias{mae}
\title{Mean Absolute Error}
\usage{
mae(y1, y2)
}
\arguments{
\item{y1}{A numeric vector.}

\item{y2}{A numeric vector.}
}
\description{
This function provides a simple way to compute MAE.
}
