% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_table}
\alias{make_table}
\title{Generate Variable Tables}
\usage{
make_table(data = NULL)
}
\arguments{
\item{data}{A data frame of labelled vectors}
}
\description{
Provided a data frame of labelled vectors, this function returns a formatted summary table.
}
