% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{party_leaders}
\alias{party_leaders}
\title{List of Labour, Conservative, and Liberal Party Leaders}
\format{
A data frame:
\describe{
\item{leader}{Name of party leader}
\item{start}{Date they became party leader}
\item{end}{Date they stood down as party leader}
\item{party}{The leader's party}
}
}
\usage{
party_leaders
}
\description{
A dataset containing all leaders of the Conservative Party, the Labour Party, and the Liberals/Liberal Democrats.
}
\keyword{datasets}
