% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pollbase}
\alias{pollbase}
\title{Raw PollBase data in long-format}
\format{
A data frame:
\describe{
\item{id}{Unique poll identification number}
\item{start}{First day of fieldwork}
\item{end}{Last day of fieldwork}
\item{pollster}{Polling company that conducted the poll}
\item{n}{Sample size}
\item{con}{Voting intention: Conservative}
\item{lab}{Voting intention: Labour}
\item{lib}{Voting intention: Liberal}
}
}
\usage{
pollbase
}
\description{
A dataset containing Mark Pack's raw pollbase figures for each of Britain's three main parties, tidied and converted to long-format for easy analysis.
}
\keyword{datasets}
