% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pollbasepro}
\alias{pollbasepro}
\title{Daily Estimates of British Voting Intention, 1955 to Present}
\format{
A data frame:
\describe{
\item{date}{Date}
\item{con_est}{Posterior mean: Conservative voting intention}
\item{con_err}{Posterior error: Conservative voting intention}
\item{lab_est}{Posterior mean: Labour voting intention}
\item{lab_err}{Posterior error: Labour voting intention}
\item{lib_est}{Posterior mean: Liberal voting intention}
\item{lib_err}{Posterior error: Liberal voting intention}
}
}
\usage{
pollbasepro
}
\description{
A dataset containing daily estimates of aggregate voting intention in Britain for each of the country's three largest parties between 1955 and the present.
}
\keyword{datasets}
