% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prime_ministers}
\alias{prime_ministers}
\title{List of British Prime Ministers}
\format{
A data frame:
\describe{
\item{prime_minister}{Name of Prime Minister}
\item{pm_party}{Prime Minister's party}
\item{start}{First day in term as Prime Minister}
\item{end}{Last day in term as Prime Minister}
}
}
\usage{
prime_ministers
}
\description{
A dataset containing the names of all British Prime Ministers.
}
\keyword{datasets}
