% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{red_wall}
\alias{red_wall}
\title{List of "Red Wall" Constituencies}
\format{
A data frame:
\describe{
\item{name}{Red wall constituency name}
\item{gss_code}{Government Statistical Service code}
}
}
\usage{
red_wall
}
\description{
A dataset containing all "red wall" constituencies targetted by the Conservatives at the 2019 election. Compiled from James Kanagasoorium's original list that spawned the term.
}
\keyword{datasets}
