% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmse}
\alias{rmse}
\title{Root-Mean-Square Error}
\usage{
rmse(y1, y2)
}
\arguments{
\item{y1}{A numeric vector.}

\item{y2}{A numeric vector.}
}
\description{
This function provides a simple way to compute RMSE.
}
