% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samplesizes}
\alias{samplesizes}
\title{Imputed Sample Sizes for British Polls}
\format{
A data frame:
\describe{
\item{date}{Date}
\item{n_est}{Estimated sample size}
}
}
\usage{
samplesizes
}
\description{
A dataset containing daily estimates of likely sample sizes for British voting intention polls.
}
\keyword{datasets}
