% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_info}
\alias{save_info}
\title{Quicky and Easy Session Info .txt Files}
\usage{
save_info(path = "session_info.txt")
}
\arguments{
\item{path}{The file path to where you would like to save your session information.}
}
\description{
This function provides a simple way to export session information.
}
