% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swing.R
\name{swing}
\alias{swing}
\title{Compute Two-Party Swing}
\usage{
swing(a1, a2, b1, b2, type = "butler")
}
\arguments{
\item{a1}{Party A vote share at time 1.}

\item{a2}{Party A vote share at time 2.}

\item{b1}{Party B vote share at time 1.}

\item{b2}{Party B vote share at time 2.}

\item{type}{Which swing calculation to use. Either "butler" or "steed". Defaults to "butler".}
}
\description{
This convenience function takes two numeric vectors of vote share data and returns the estimated two-party swing.
}
\examples{
\dontrun{
swing(a1 = .5, a2 = .4, b1 = .5, b2 = .6, type = "butler")
}
}
